Initial access token check, redirect to login screen or normal activity screen based on results.
- errors (via plist) included, but not utilized from RA mapping right now.
- Storyboard default initial controller removed, set pragmatically now.
- csv to plist converter for use in a simple csv added
- other code focused on handling the login/auth and completion process.

+++ quickdialog/QDateTimeInlineElement.m
@@ -137,7 +137,7 @@
     NSInteger seconds = ti % 60;
     NSInteger minutes = (ti / 60) % 60;
     NSInteger hours = (ti / 3600);
-    return [NSString stringWithFormat:@"%02i:%02i:%02i", hours, minutes, seconds];
+    return [NSString stringWithFormat:@"%02i:%02i:%02i", (int)hours, (int)minutes, (int)seconds];
 }
